requirements.txt
autopep8==1.5.4
pdf2image==1.14.0
Pillow==7.2.0
pycodestyle==2.6.0
pyocr==0.7.2
toml==0.10.1
